Combine WildcardMatcher and WildcardsList
