+++ lib/licensed.rb
@@ -5,8 +5,7 @@ require "licensed/dependency_record"
 require "licensed/dependency"
 require "licensed/git"
 require "licensed/sources"
-require "licensed/wildcard_matcher"
-require "licensed/matchers_list"
+require "licensed/wildcards_list"
 require "licensed/configuration"
 require "licensed/reporters"
 require "licensed/commands"
